requirements.txt
# --- Core Orchestration & Graph ---
langgraph>=0.2.0
langgraph-checkpoint-sqlite
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-neo4j>=0.2.0
aiosqlite

# --- Web Server & API ---
fastapi>=0.110.0
uvicorn[standard]
websockets
python-multipart

# --- AI & LLM Clients ---
openai>=1.50.0          # Client for vLLM
pydantic>=2.9.0
pydantic-settings>=2.5.0
httpx>=0.27.0           # Pinning this prevents the loop!

# --- Observability (The "Eye") ---
arize-phoenix>=4.0.0
openinference-instrumentation-langchain>=0.1.0
opentelemetry-sdk>=1.26.0
opentelemetry-exporter-otlp>=1.26.0
opentelemetry-instrumentation-fastapi>=0.47b0

# --- Memory (The "Context") ---
mem0ai>=0.1.20
qdrant-client>=1.12.0
sentence-transformers>=3.0.0
# kuzu                  # Comment out if not strictly needed immediately
rank-bm25

# --- Database ---
sqlalchemy>=2.0.0
asyncpg>=0.29.0
greenlet

# --- Skill Library (Code Storage) ---
chromadb>=0.5.5         # Semantic code snippet library

# --- Tools & Sandbox (The "Body") ---
psutil

# --- Audio ---
faster-whisper
numpy<2.0.0             # Critical: faster-whisper breaks on numpy 2.0
kokoro-onnx
soundfile

# --- Utilities ---
python-dotenv
tavily-python
yfinance

# --- Testing ---
pytest
pytest-asyncio

